+++ Sources/NeoSwift/protocol/core/polling/BlockIndexPolling.swift
@@ -17,18 +17,19 @@ public struct BlockIndexPolling {
     var currentBlockIndex = BlockIndexActor()
     
     public func blockIndexPublisher(_ neoSwift: NeoSwift, _ executor: DispatchQueue, _ pollingInterval: Int) -> AnyPublisher<Int, Error> {
-        return Timer.publish(every: Double(pollingInterval) / 1000, on: .current, in: .default)
+        return Timer.publish(every: Double(pollingInterval) / 1000, on: .main, in: .common)
             .autoconnect()
             .setFailureType(to: Error.self)
             .asyncMap { _ -> [Int]? in
                 let latestBlockIndex = try await neoSwift.getBlockCount().send().getResult() - 1
                 if await currentBlockIndex.blockIndex == nil {
                     await currentBlockIndex.setIndex(latestBlockIndex)
+                    return [latestBlockIndex]
                 }
-                if await latestBlockIndex > currentBlockIndex.blockIndex! {
-                    let currIndex = await currentBlockIndex.blockIndex!
-                    await currentBlockIndex.setIndex(latestBlockIndex )
-                    return Array((currIndex + 1)...latestBlockIndex)
+                let currentIdx = await currentBlockIndex.blockIndex ?? 0
+                if latestBlockIndex > currentIdx {
+                    await currentBlockIndex.setIndex(latestBlockIndex)
+                    return Array((currentIdx + 1)...latestBlockIndex)
                 }
                 return nil
             }